services/sentiment-stream/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.8.2
huggingface-hub==0.11.1
idna==3.4
kafka-python==2.0.2
numpy==1.23.5
packaging==22.0
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scipy==1.9.3
six==1.16.0
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
